requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.7.2
django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
pydotplus==2.0.2
pyparsing==3.1.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
